requirements.txt
requests==2.26.0
pytest==6.2.5
python-dotenv==0.19.2
pandas==1.3.5
